06-orchestrator/requirements.txt
apache-airflow-providers-apache-spark==5.3.0    # 4.10.0
apache-airflow-providers-amazon==9.9.0          # 8.28.0
pyspark==3.5.5
# pyiceberg[pyarrow]==0.7.0
# pandas==2.2.0
# pyarrow==15.0.0
# protobuf==6.31.1  # 4.25.0
# PyPDF2==3.0.1
# pdfplumber==0.10.0